chore: move GlobalConfig declaration to config.ts
